+++ src/app/app.component.spec.ts
@@ -3,15 +3,11 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule
-      ],
-      declarations: [
-        AppComponent
-      ],
-    }).compileComponents();
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [ RouterTestingModule ],
+      declarations: [ AppComponent ],
+    });
   });
 
   it('should create the app', () => {
@@ -26,10 +22,10 @@ describe('AppComponent', () => {
     expect(app.title).toEqual('shop');
   });
 
-  it('should render title', () => {
+  it('should render appTitle', () => {
     const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
     fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('.content span')?.textContent).toContain('shop app is running!');
+    expect(app.appTitle.nativeElement.textContent).toContain('Shop App');
   });
 });