+++ src/app/orders/validators/custom.validator.ts
@@ -2,7 +2,7 @@ import { AbstractControl, ValidationErrors } from "@angular/forms";
 
 export class CustomValidators {
     static startCase({ value }: AbstractControl): ValidationErrors | null {
-        const isStartCase = /^[A-Z]/.test(value);
+        const isStartCase = /^\p{Lu}\p{Ll}+$/.test(value);
         if (value !== '' && !isStartCase) {
             return { startCase: true }
         }